Unmatched parallelism enabled by aces for IBO creation ( ͡° ͜ʖ ͡°)

+++ programs/dark-bonds/src/state/lockup.rs
@@ -8,7 +8,7 @@ pub struct Lockup {
     /** Bump.*/
     pub bump: u8,
     /** Index of this lockup.*/
-    pub index: u32,
+    pub index: u16,
     /** Minimum lockup period in seconds.*/
     pub period: i64,
     /** Yearly APY for this lockup. As it is f64 disguised as u64 due to solita, need to divide by 1000*/
